getWorksetRunIds {openMpp}	R Documentation

Return ids of model run results for specified working set

Description

Return ids of model run results where input parameters are from specified working set.

Usage

getWorksetRunIds(dbCon, worksetId)

Arguments

dbCon
database connection

worksetId
id of parameters working set, must be positive integer

Details

Return ids of model run results where input parameters are from specified working set.

Please keep in mind: there is no established link in database between input data working set and model run results. Input data can be modified or even deleted after model run. If you really want to find input parameters values for specific model run then selectRunParameter should be used.

It is recommended to create modeling task by createTask to include multiple working sets of input parameters and run the model using such input.

Value

Data frame with integer $run_id column

Note

To run examples you must have modelOne database modelOne.sqlite in current directory

Author(s)

amc1999

References

OpenM++ documentation: http://ompp.sourceforge.net/wiki/

See Also

createTask getModel getFirstRunId getLastRunId getDefaultWorksetId getWorksetIdByName selectRunParameter

Examples

  theDb <- dbConnect(RSQLite::SQLite(), "modelOne.sqlite", synchronous = "full")
  invisible(dbGetQuery(theDb, "PRAGMA busy_timeout = 86400")) # recommended

  runIdRs <- getWorksetRunIds(theDb, 2L)

  dbDisconnect(theDb)

[Package openMpp version 0.8.1 Index]