OpenM++ Model Worksets

Functions for creating, copying, merging, retrieving, and deleting worksets.

Usage

get_workset(model, set)

get_worksets_list(model)

get_worksets(model)

get_scenarios(model)

get_workset_status(model, set)

get_workset_status_default(model)

set_workset_readonly(model, set, readonly)

create_workset(data)

merge_workset(workset)

update_workset_param_csv(workset, csv)

replace_workset(workset)

delete_workset(model, set)

delete_workset_param(model, set, name)

update_workset_param(model, set, name, data)

copy_param_run_to_workset(model, set, name, run)

merge_param_run_to_workset(model, set, name, run)

copy_param_workset_to_workset(model, set, name, from)

merge_param_workset_to_workset(model, set, name, from)

upload_workset_params(model, set, data)

delete_scenario(model, set)

Arguments

model

Model digest or model name.

set

Name of workset (input set of model parameters).

readonly

Boolean. Should workset be read-only?

data

Data used for the body of the request.

workset

Workset metadata.

csv

CSV file path.

name

Output table name.

run

Model run digest, run stamp or run name, modeling task run stamp or task run name.

from

Source workset name.

Value

A list, tibble, or nothing (invisibly).