openMpp-package {openMpp}	R Documentation

Read and write OpenM++ database from R

Description

Set input parameters and get output results of OpenM++ models using R.

Details

Package: 	 openMpp

Type: 	 Package

Version: 	 0.8

Date: 	 2016-12-27

License: 	 MIT license

createTask

Create new modeling task

createWorkset

Create new working set of model parameters

createWorksetBasedOnRun

Create new working set of model parameters based on parameters from existing model run

copyWorksetParameterFromRun

Copy parameters to working set from existing model run

getDefaultWorksetId

Return default working set id

getFirstRunId

Return id of first model run results

getLastRunId

Return id of last model run results

getTaskFirstRunId

Return first id of modeling task run

getTaskLastRunId

Return last id of modeling task run

getLanguages

Return language rows from lang_lst table

getModel

Find OpenM++ model definition in database

getTaskIdByName

Return modeling task id by name

getWorksetIdByName

Return working set id by name

getWorksetRunIds

Return ids of model run results for specified working set

selectRunAccumulator

Select output table accumulator values from model run result

selectRunOutputValue

Select output table expression values from model run result

selectRunParameter

Select parameter values from model run result

selectTask

Select modeling task text (name, description, notes) and input workset id's

selectTaskList

Select list of modeling tasks for the model

selectTaskRun

Select modeling task run: status and results

selectTaskRunList

Select list of modeling task runs

setReadonlyDefaultWorkset

Set or clear read-only status of default working set

setReadonlyWorkset

Set or clear read-only status for default working set of model parameters

setTaskWaitCompleted

Set modeling task "wait completed" status

updateTask

Update modeling task with new text (name, description, notes) or additional input working sets

updateWorksetParameter

Update parameters working set with new values and value notes

toQuoted

Return SQL-quoted string: O'Connor -> 'O”Connor'

Author(s)

amc1999

Maintainer: amc1999 <openmpp99@gmail.com>

References

OpenM++ documentation: http://ompp.sourceforge.net/wiki/

See Also

createTask createWorkset createWorksetBasedOnRun copyWorksetParameterFromRun getDefaultWorksetId getFirstRunId getLastRunId getLanguages getModel getTaskIdByName getTaskFirstRunId getTaskLastRunId getWorksetIdByName getWorksetRunIds selectRunAccumulator selectRunOutputValue selectRunParameter selectTask selectTaskList selectTaskRun selectTaskRunList setReadonlyDefaultWorkset setReadonlyWorkset setTaskWaitCompleted updateTask updateWorksetParameter toQuoted

[Package openMpp version 0.8.3 Index]