selectRunParameter {openMpp}	R Documentation

Select parameter values from model run result

Description

Select parameter values from model run result

Usage

selectRunParameter(dbCon, defRs, runId, paramName)

Arguments

dbCon
database connection

defRs
model definition: database rows describing model input parameters and output tables

runId
id of model run results, must be positive integer

paramName
parameter name, i.e.: "sexAge"

Details

That call return values of input parameter from model run results with id runId.

You must use getModel function in order to find model definition defRs.

Value

Data frame of database rows with parameter dimensions and values:

sub_id
parameter sub-value id or zero if parameter has no sub-values

dim0,...,dimN
dimension items enum ids (not returned if rank is zero)

value
parameter value

Note

To run examples you must have modelOne database modelOne.sqlite in current directory

Author(s)

amc1999

References

OpenM++ documentation: http://ompp.sourceforge.net/wiki/

See Also

getModel getFirstRunId getLastRunId selectRunAccumulator selectRunOutputValue

Examples

  #
  # connect to database
  #
  theDb <- dbConnect(RSQLite::SQLite(), "modelOne.sqlite", synchronous = "full")
  invisible(dbGetQuery(theDb, "PRAGMA busy_timeout = 86400")) # recommended
  
  # get definition for "modelOne" model
  #
  # use model name only to identify the model
  # if there are multiple versions of the model in database 
  # then first one is selected: where model id = min(model_id) for "modelOne"
  #
  defRs <- getModel(theDb, "modelOne")
  
  # get first run id of that model
  runId <- getFirstRunId(theDb, defRs)
  if (runId <= 0L) stop("model run results not found")
  
  # select parameter "ageSex" value from model run results
  #  
  paramValueRs <- selectRunParameter(theDb, defRs, runId, "ageSex")
  
  dbDisconnect(theDb)

[Package openMpp version 0.8.3 Index]