setTaskWaitCompleted {openMpp}	R Documentation

Set modeling task "wait completed" status

Description

Set modeling task "wait completed" status

Usage

setTaskWaitCompleted(dbCon, taskRunId, isWaitCompleted = FALSE)

Arguments

dbCon
database connection

taskRunId
id of modeling task run

isWaitCompleted
if true then signal to running model "task is ready to be completed"

Details

Using this setTaskWaitCompleted call to signal currently running model "task is ready to be completed".

Task run status can be one of:

i = not yet started

p = run in progress

w = run in progress, wait for additional input

s = completed successfully

e = failed (error)

x = reserved

Status w means task can be dynamically updated by external script. Model executable is waiting for additional input into or for "ready to be completed" signal. Use setTaskWaitCompleted function to mark task as "ready to be completed". Use updateTask function to insert additional task input.

Model executable must be running with -OpenM.TaskWait true argument in order to handle dynamical task input. For example:

modelOne -OpenM.TaskName taskOne -OpenM.TaskWait true

Value

Return task id or 0L on error

Note

To run examples you must have modelOne database modelOne.sqlite in current directory

Author(s)

amc1999

References

OpenM++ documentation: http://ompp.sourceforge.net/wiki/

See Also

getModel getTaskIdByName createTask

[Package openMpp version 0.8.3 Index]